Apply schema grammar fixes for JSON path array references
Illuminate\Database\Query\Grammars\Grammar@wrapJsonFieldAndPath() and
Illuminate\Database\Schema\Grammars\Grammar@wrapJsonFieldAndPath() were
copy and pasted duplicates but the query version had JSON path fixes
applied that the schema version doesn't have.

Instead add a trait to make the two classes share these methods.
